+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to this project are documented in this file.
 
 - [BREAKING] dropping the LayerId does not cause the layer to be removed. Only if the layer is also empty does that happen now.
 - [BREAKING] LayerId is now !Send.
+- [BREAKING] Context::new() stopped taking the mouse location. There is no way (yet) to get it at the location that you want to create a context.
 - Remove unesesary lifetime parameter in Context.
 - Implement clone for LayerId.